refactor: share unocss config across multiple packages
