+++ cspell.config.yaml
@@ -67,6 +67,7 @@ words:
   - gpuu
   - grammyjs
   - groq
+  - gugi
   - guiiai
   - hfspaceup
   - hfsup
@@ -138,6 +139,7 @@ words:
   - pthread
   - qwen
   - rehype
+  - reka
   - rushstack
   - Saccade
   - saccades
