+++ package.json
@@ -36,6 +36,7 @@
   "devDependencies": {
     "@antfu/eslint-config": "^4.12.0",
     "@antfu/ni": "^24.3.0",
+    "@iconify/utils": "^2.3.0",
     "@types/node": "^22.14.1",
     "@unocss/eslint-config": "^66.1.0-beta.12",
     "@unocss/eslint-plugin": "^66.1.0-beta.12",
@@ -51,6 +52,7 @@
     "typescript": "~5.8.3",
     "unbuild": "^3.5.0",
     "unocss": "^66.1.0-beta.12",
+    "unocss-preset-scrollbar": "^3.2.0",
     "vite": "^6.3.3",
     "vite-plugin-inspect": "^11.0.1",
     "vitest": "^3.1.2"
