+++ packages/stage-ui/src/components/Menu/RadioCardManySelect.vue
@@ -106,7 +106,7 @@ function updateCustomValue(value: string) {
       <div class="relative">
         <!-- Horizontally scrollable container -->
         <div
-          class="scrollbar-hide grid auto-cols-[350px] grid-flow-col max-h-[calc(100dvh-7lh)] gap-4 overflow-x-auto pb-4"
+          class="grid auto-cols-[350px] grid-flow-col max-h-[calc(100dvh-7lh)] gap-4 overflow-x-auto pb-4 scrollbar-none"
           :class="[
             isListExpanded ? 'grid-cols-1 md:grid-cols-2 grid-flow-row auto-cols-auto' : '',
           ]"
